chore: port Droppable to Typescript
