--- src/Droppable/DroppableEvent/index.js
@@ -1,1 +1,0 @@
-export * from './DroppableEvent';
